fixed a compiler warning by providing an int type of the argument to printf, as specified by the format.

+++ examples/sky/test-coffee.c
@@ -317,7 +317,7 @@ PROCESS_THREAD(testcoffee_process, ev, data)
   printf("Coffee format: %d\n", cfs_coffee_format());
   printf("Coffee file test: %d\n", coffee_file_test());
   printf("Coffee garbage collection test: %d\n", coffee_gc_test());
-  printf("Test time: %d seconds\n", clock_seconds() - start);
+  printf("Test time: %d seconds\n", (int)(clock_seconds() - start));
 
   PROCESS_END();
 }
